requirements.txt
streamlit==1.44.0
opencv-python==4.11.0.86
numpy==2.2.4
matplotlib==3.10.1
pillow==11.1.0